+++ libs/@guardian/source-foundations/.storybook/preview.tsx
@@ -1,7 +1,0 @@
-export const parameters = {
-	options: {
-		storySort: {
-			method: 'alphabetical',
-		},
-	},
-};