+++ libs/@guardian/source-foundations/rollup.config.js
@@ -1,1 +1,0 @@
-export * as default from '../../../configs/rollup.config.js';