+++ libs/@guardian/source/src/foundations/typography/stories/typography.stories.tsx
@@ -1,0 +1,20 @@
+import { TypographyPresets } from './TypographyPresets';
+
+export default {
+	title: 'Foundations/Typography',
+
+	parameters: {
+		previewTabs: {
+			canvas: {
+				hidden: true,
+			},
+		},
+
+		viewMode: 'docs',
+	},
+};
+
+export const Presets = {
+	render: () => <TypographyPresets />,
+	name: 'presets',
+};